Convert to a single class, remove facade, and add base class for both netsnmp classes
