+++ LibreNMS/Data/Source/NetSnmpTranslate.php
@@ -1,7 +1,7 @@
 <?php
 
 /*
- * SNMP.php
+ * NetSnmpTranslate.php
  *
  * -Description-
  *
@@ -20,8 +20,8 @@
  *
  * @package    LibreNMS
  * @link       http://librenms.org
- * @copyright  2021 Tony Murray
- * @author     Tony Murray <murraytony@gmail.com>
+ * @copyright  2026 Steven Wilton
+ * @author     Steven Wilton <swilton@fluentit.au>
  */
 
 namespace LibreNMS\Data\Source;
@@ -38,13 +38,13 @@ use LibreNMS\Util\Oid;
 use Log;
 use Symfony\Component\Process\Process;
 
-class NetSnmpTranslate implements SnmpTranslateInterface
+class NetSnmpTranslate extends NetSnmpCmd
 {
-    private array $options = [];
-    private array $mibDirs = [];
-    private Device $device;
+    protected array $options = [];
+    protected array $mibDirs = [];
+    protected Device $device;
     // defaults for net-snmp https://net-snmp.sourceforge.io/docs/man/snmpcmd.html
-    private array $mibs = ['SNMPv2-TC', 'SNMPv2-MIB', 'IF-MIB', 'IP-MIB', 'TCP-MIB', 'UDP-MIB', 'NET-SNMP-VACM-MIB'];
+    protected array $mibs = ['SNMPv2-TC', 'SNMPv2-MIB', 'IF-MIB', 'IP-MIB', 'TCP-MIB', 'UDP-MIB', 'NET-SNMP-VACM-MIB'];
 
     public function __construct()
     {
@@ -54,7 +54,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
     /**
      * Easy way to start a new instance
      */
-    public static function make(): SnmpTranslateInterface
+    public static function make(): NetSnmpTranslate
     {
         return new static;
     }
@@ -63,7 +63,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
      * Specify a device to make the snmp query against.
      * By default the query will use the primary device.
      */
-    public function device(Device $device): SnmpTranslateInterface
+    public function device(Device $device): NetSnmpTranslate
     {
         $this->device = $device;
 
@@ -74,7 +74,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
      * Set an additional MIB directory to search for MIBs.
      * You do not need to specify the base and os directories, they are already included.
      */
-    public function mibDir(?string $dir): SnmpTranslateInterface
+    public function mibDir(?string $dir): NetSnmpTranslate
     {
         $this->mibDirs[] = $dir;
 
@@ -85,7 +85,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
      * Set MIBs to use for this query. Base mibs are included by default.
      * They will be appended to existing mibs unless $append is set to false.
      */
-    public function mibs(array $mibs, bool $append = true): SnmpTranslateInterface
+    public function mibs(array $mibs, bool $append = true): NetSnmpTranslate
     {
         $this->mibs = $append ? array_merge($this->mibs, $mibs) : $mibs;
 
@@ -95,7 +95,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
     /**
      * Output all OIDs numerically
      */
-    public function numeric(bool $numeric = true): SnmpTranslateInterface
+    public function numeric(bool $numeric = true): NetSnmpTranslate
     {
         $this->options = $numeric
             ? array_merge($this->options, ['-On'])
@@ -107,7 +107,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
     /**
      * Hide MIB in output
      */
-    public function hideMib(): SnmpTranslateInterface
+    public function hideMib(): NetSnmpTranslate
     {
         $this->options = array_merge($this->options, ['-Os']);
 
@@ -139,7 +139,7 @@ class NetSnmpTranslate implements SnmpTranslateInterface
         return $this->exec('snmptranslate', [$oid])->value();
     }
 
-    private function buildCli(string $command, array $oids): array
+    protected function buildCli(string $command, array $oids): array
     {
         $cmd = [LibrenmsConfig::get($command, $command)];
 
@@ -147,88 +147,5 @@ class NetSnmpTranslate implements SnmpTranslateInterface
         array_push($cmd, '-m', implode(':', $this->mibs));
 
         return array_merge($cmd, $this->options, $oids);
-    }
-
-    private function exec(string $command, array $oids): SnmpResponse
-    {
-        $measure = Measurement::start($command);
-        $proc = new Process($this->buildCli($command, $oids));
-        $proc->setTimeout(LibrenmsConfig::get('snmp.exec_timeout', 1200));
-
-        $this->logCommand($proc->getCommandLine());
-
-        $proc->run();
-        $exitCode = $proc->getExitCode();
-        $output = $proc->getOutput();
-        $stderr = $proc->getErrorOutput();
-
-        // check exit code and log possible bad auth
-        $this->checkExitCode($exitCode, $stderr);
-        $this->logOutput($output, $stderr);
-
-        $measure->manager()->recordSnmp($measure->end());
-
-        return new SnmpResponse($output, $stderr, $exitCode);
-    }
-
-    private function mibDirectories(): string
-    {
-        $base = LibrenmsConfig::get('mib_dir');
-        $dirs = [$base];
-
-        // os group
-        if ($os_group = LibrenmsConfig::getOsSetting($this->device->os, 'group')) {
-            if (file_exists("$base/$os_group")) {
-                $dirs[] = "$base/$os_group";
-            }
-        }
-
-        // os directory
-        $os_mibdir = LibrenmsConfig::getOsSetting($this->device->os, 'mib_dir');
-        if ($os_mibdir && is_string($os_mibdir)) {
-            $dirs[] = "$base/$os_mibdir";
-        } elseif (file_exists($base . '/' . $this->device->os)) {
-            $dirs[] = $base . '/' . $this->device->os;
-        }
-
-        foreach ($this->mibDirs as $mibDir) {
-            $dirs[] = "$base/$mibDir";
-        }
-
-        // remove trailing /, remove empty dirs, and remove duplicates
-        $dirs = array_unique(array_filter(array_map(fn ($dir) => rtrim((string) $dir, '/'), $dirs)));
-
-        return implode(':', $dirs);
-    }
-
-    private function checkExitCode(int $code, string $error): void
-    {
-        if ($code) {
-            if (Str::startsWith($error, 'Invalid authentication protocol specified')) {
-                Eventlog::log('Unsupported SNMP authentication algorithm - ' . $code, $this->device, 'poller', Severity::Error);
-            } elseif (Str::startsWith($error, 'Invalid privacy protocol specified')) {
-                Eventlog::log('Unsupported SNMP privacy algorithm - ' . $code, $this->device, 'poller', Severity::Error);
-            }
-            Log::debug('Exitcode: ' . $code, [$error]);
-        }
-    }
-
-    private function logCommand(string $command): void
-    {
-        if (Debug::isEnabled() && ! Debug::isVerbose()) {
-            Log::debug('SNMP[%c' . $command . '%n]', ['color' => true]);
-        } elseif (Debug::isVerbose()) {
-            Log::debug('SNMP[%c' . $command . '%n]', ['color' => true]);
-        }
-    }
-
-    private function logOutput(string $output, string $error): void
-    {
-        if (Debug::isEnabled() && ! Debug::isVerbose()) {
-            Log::debug($output);
-        } elseif (Debug::isVerbose()) {
-            Log::debug($output);
-        }
-        Log::debug($error);
     }
 }
