+++ LibreNMS/OS/Awplus.php
@@ -31,11 +31,11 @@ use App\Models\Device;
 use App\Models\Transceiver;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
+use LibreNMS\Data\Sources\NetSnmpTranslate;
 use LibreNMS\Interfaces\Discovery\OSDiscovery;
 use LibreNMS\Interfaces\Discovery\TransceiverDiscovery;
 use LibreNMS\OS;
 use SnmpQuery;
-use SnmpTranslate;
 
 class Awplus extends OS implements OSDiscovery, TransceiverDiscovery
 {
@@ -52,7 +52,7 @@ class Awplus extends OS implements OSDiscovery, TransceiverDiscovery
         //Instead use sysObjectID.0
 
         if (Str::contains($hardware, 'SBx81')) {
-            $hardware = SnmpTranslate::hideMib()->mibs(['AT-PRODUCT-MIB'])->translate($device->sysObjectID);
+            $hardware = NetSnmpTranslate::make()->hideMib()->mibs(['AT-PRODUCT-MIB'])->translate($device->sysObjectID);
             $hardware = str_replace('at', 'AT-', $hardware);
 
             // Features and Serial is set to Controller card 1.5 or 1.6
