+++ LibreNMS/OS/Axos.php
@@ -31,6 +31,7 @@ use App\Models\Device;
 use App\Models\EntPhysical;
 use App\Models\Transceiver;
 use Illuminate\Support\Collection;
+use LibreNMS\Data\Sources\NetSnmpTranslate;
 use LibreNMS\Interfaces\Discovery\OSDiscovery;
 use LibreNMS\Interfaces\Discovery\TransceiverDiscovery;
 use LibreNMS\OS;
@@ -75,7 +76,7 @@ class Axos extends OS implements OSDiscovery, TransceiverDiscovery
         $inventory = new Collection;
         $physical_index = 1;
 
-        $physical_name = \SnmpTranslate::hideMib()->mibs(['CALIX-PRODUCT-MIB'])->translate($this->getDevice()->sysObjectID);
+        $physical_name = NetSnmpTranslate::make->hideMib()->mibs(['CALIX-PRODUCT-MIB'])->translate($this->getDevice()->sysObjectID);
         $serial_number = \SnmpQuery::get('Axos-System-MIB::axosSystemChassisSerialNumber.0')->value();
         $inventory->push(new EntPhysical([
             'entPhysicalIndex' => $physical_index++,
