+++ LibreNMS/OS/Junose.php
@@ -27,6 +27,7 @@
 namespace LibreNMS\OS;
 
 use App\Models\Device;
+use LibreNMS\Data\Sources\NetSnmpTranslate;
 
 class Junose extends \LibreNMS\OS
 {
@@ -38,7 +39,7 @@ class Junose extends \LibreNMS\OS
             return;
         }
 
-        $junose_hardware = \SnmpTranslate::mibs(['Juniper-Products-MIB'])->translate($device->sysObjectID);
+        $junose_hardware = NetSnmpTranslate::make()->mibs(['Juniper-Products-MIB'])->translate($device->sysObjectID);
         $device->hardware = $this->rewriteHardware($junose_hardware) ?: null;
 
         $junose_version = \SnmpQuery::get('Juniper-System-MIB::juniSystemSwVersion.0')->value();
