+++ LibreNMS/OS/Traits/YamlOSDiscovery.php
@@ -30,10 +30,10 @@ use App\Models\Device;
 use App\Models\Location;
 use App\View\SimpleTemplate;
 use Illuminate\Support\Arr;
+use LibreNMS\Data\Sources\NetSnmpTranslate;
 use LibreNMS\Util\Oid;
 use LibreNMS\Util\StringHelpers;
 use Log;
-use SnmpTranslate;
 
 trait YamlOSDiscovery
 {
@@ -62,7 +62,7 @@ trait YamlOSDiscovery
         }
 
         if (isset($os_yaml['hardware_mib'])) {
-            $device->hardware = SnmpTranslate::mibs([$os_yaml['hardware_mib']])->hideMib()->translate($device->sysObjectID);
+            $device->hardware = NetSnmpTranslate::make()->mibs([$os_yaml['hardware_mib']])->hideMib()->translate($device->sysObjectID);
 
             if (! empty($os_yaml['hardware_regex'])) {
                 $this->parseRegex($os_yaml['hardware_regex'], $device->hardware);
