+++ app/Console/Commands/SnmpTranslate.php
@@ -5,6 +5,7 @@ namespace App\Console\Commands;
 use App\Facades\LibrenmsConfig;
 use App\Models\Device;
 use Illuminate\Support\Collection;
+use LibreNMS\Data\Source\NetSnmpTranslate;
 use LibreNMS\Data\Source\SnmpResponse;
 
 class SnmpTranslate extends SnmpFetch
@@ -51,8 +52,8 @@ class SnmpTranslate extends SnmpFetch
         $res = new SnmpResponse('');
         // translate does not support multiple oids (should it?)
         foreach ($this->oids as $oid) {
-            $textual = \SnmpTranslate::numeric(false)->device($device)->mibs(['ALL'])->translate($oid);
-            $numeric = \SnmpTranslate::numeric(true)->device($device)->mibs(['ALL'])->translate($oid);
+            $textual = NetSnmpTranslate::make()->numeric(false)->device($device)->mibs(['ALL'])->translate($oid);
+            $numeric = NetSnmpTranslate::make()->numeric(true)->device($device)->mibs(['ALL'])->translate($oid);
 
             $response = new SnmpResponse("$textual = $numeric\n");
             $res = $res->append($response);
