--- app/Facades/FacadeAccessorSnmpTranslate.php
@@ -1,41 +1,0 @@
-<?php
-
-/*
- * FacadeAccessorSnmp.php
- *
- * -Description-
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @package    LibreNMS
- * @link       http://librenms.org
- * @copyright  2026 Steven Wilton
- * @author     Steven Wilton <swilton@fluentit.au>
- */
-
-namespace App\Facades;
-
-use Illuminate\Support\Facades\Facade;
-use LibreNMS\Data\Source\NetSnmpTranslate;
-
-class FacadeAccessorSnmpTranslate extends Facade
-{
-    protected static function getFacadeAccessor()
-    {
-        // always resolve a new instance
-        self::clearResolvedInstance(NetSnmpTranslate::class);
-
-        return NetSnmpTranslate::class;
-    }
-}
