+++ config/app.php
@@ -23,7 +23,6 @@ return [
         'PluginManager' => App\Facades\PluginManager::class,
         'Rrd' => App\Facades\Rrd::class,
         'SnmpQuery' => App\Facades\FacadeAccessorSnmp::class,
-        'SnmpTranslate' => App\Facades\FacadeAccessorSnmpTranslate::class,
         'LibrenmsConfig' => App\Facades\LibrenmsConfig::class,
     ])->forget([
         'Http', // don't use Laravel Http facade, LibreNMS has its own wrapper
